requirements.txt
fastapi==0.110.3
numpy==1.26.4
apscheduler==3.10.4
uvicorn==0.27.1
openai==1.16.1
mysql-connector-python==8.3.0
pandas==2.2.1
python-dotenv==1.0.1
torch==2.2.1
transformers==4.38.2
